# codex-plus-contracts/api/client-openapi.yaml
openapi: 3.1.0
info:
  title: Codex++ Client API
  version: 0.1.0-mvp
  description: >
    MVP contract for Codex++ desktop client APIs. The desktop client consumes
    aggregated runtime snapshots only; pricing, model policy, quota policy and
    entitlement rules remain server-side. Clients must not hardcode prices,
    plan logic, model billing multipliers, quota thresholds, rate-limit policy,
    or renewal/purchase copy; these are displayed or acted on only from the
    bootstrap/config snapshot.
servers:
  - url: https://api.codex-plus.example
    description: Production placeholder
  - url: http://localhost:3000
    description: Local development placeholder
security:
  - bearerAuth: []
paths:
  /api/v1/auth/desktop/start:
    post:
      operationId: startDesktopBrowserLogin
      summary: Create a browser-approved desktop login session.
      description: >
        Public desktop entry point. It returns a browser authorize URL and a
        desktop-private poll token. The browser must never receive poll_token.
      security: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DesktopLoginStartRequest"
      responses:
        "200":
          description: Pending desktop browser handoff.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DesktopLoginStartEnvelope"
        "500":
          $ref: "#/components/responses/DesktopLoginUnavailable"
  /api/v1/auth/desktop/complete:
    post:
      operationId: completeDesktopBrowserLogin
      summary: Approve a desktop login from an authenticated browser session.
      description: >
        Requires the normal Web JWT. This endpoint binds the pending desktop
        session to the current browser user and returns only a completion
        status, never a desktop access token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DesktopLoginCompleteRequest"
      responses:
        "200":
          description: Desktop login approved.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DesktopLoginCompleteEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/DesktopLoginTargetMismatch"
  /api/v1/auth/desktop/poll:
    post:
      operationId: pollDesktopBrowserLogin
      summary: Let the desktop poll for browser login completion.
      description: >
        Public desktop polling endpoint. It requires both session_token and the
        desktop-private poll_token. A completed poll is one-time and consumes
        the pending session before issuing desktop tokens.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DesktopLoginPollRequest"
      responses:
        "200":
          description: Pending or completed desktop browser handoff.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DesktopLoginPollEnvelope"
        "401":
          $ref: "#/components/responses/PendingAuthUnauthorized"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    Unauthorized:
      description: Missing, expired or invalid user JWT.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          examples:
            notAuthenticated:
              value:
                code: 401
                status: error
                message: Please sign in again.
                reason: CLIENT_AUTH_NOT_AUTHENTICATED
                data: null
                error_code: CLIENT_AUTH_NOT_AUTHENTICATED
    PendingAuthUnauthorized:
      description: Missing, expired, consumed or mismatched desktop pending auth session.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          examples:
            expired:
              value:
                code: 401
                status: error
                message: pending auth session has expired
                reason: PENDING_AUTH_SESSION_EXPIRED
                data: null
                error_code: PENDING_AUTH_SESSION_EXPIRED
            browserMismatch:
              value:
                code: 401
                status: error
                message: pending auth completion code does not match this browser session
                reason: PENDING_AUTH_BROWSER_MISMATCH
                data: null
                error_code: PENDING_AUTH_BROWSER_MISMATCH
    DesktopLoginUnavailable:
      description: Desktop browser handoff could not be created.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          examples:
            createFailed:
              value:
                code: 500
                status: error
                message: failed to create desktop login session
                reason: CLIENT_AUTH_DESKTOP_SESSION_CREATE_FAILED
                data: null
                error_code: CLIENT_AUTH_DESKTOP_SESSION_CREATE_FAILED
    DesktopLoginTargetMismatch:
      description: Pending desktop session was already approved by another user.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          examples:
            mismatch:
              value:
                code: 409
                status: error
                message: desktop login session has already been approved by another user
                reason: CLIENT_AUTH_DESKTOP_TARGET_MISMATCH
                data: null
                error_code: CLIENT_AUTH_DESKTOP_TARGET_MISMATCH
  schemas:
    EnvelopeBase:
      type: object
      description: >
        All MVP responses and fixtures use a dual-compatible envelope:
        code/status/message/data/error_code/reason. Success responses use
        code=0, status=success, error_code=null and reason=null. Error
        responses use status=error; reason must equal error_code for legacy
        Sub2API consumers. During migration, manual consumers may also unwrap
        older code/message/reason-only responses.
      required: [code, status, message, reason, error_code]
      properties:
        code:
          type: integer
          description: Legacy Sub2API response code. A value of 0 means success; errors mirror the HTTP status or legacy code.
        status:
          type: string
          enum: [success, error]
          description: Contract fixture status. New fixtures must always include it.
        message:
          type: string
        reason:
          type: string
          nullable: true
          description: Legacy Sub2API error reason; equivalent to error_code for implementation checks.
        error_code:
          type: string
          nullable: true
    ErrorEnvelope:
      allOf:
        - $ref: "#/components/schemas/EnvelopeBase"
        - type: object
          required: [data]
          properties:
            data:
              type: "null"
...
    DesktopLoginStartEnvelope:
      allOf:
        - $ref: "#/components/schemas/EnvelopeBase"
        - type: object
          required: [data]
          properties:
            data:
              $ref: "#/components/schemas/DesktopLoginStartResult"
    DesktopLoginCompleteEnvelope:
      allOf:
        - $ref: "#/components/schemas/EnvelopeBase"
        - type: object
          required: [data]
          properties:
            data:
              $ref: "#/components/schemas/DesktopLoginCompleteResult"
    DesktopLoginPollEnvelope:
      allOf:
        - $ref: "#/components/schemas/EnvelopeBase"
        - type: object
          required: [data]
          properties:
            data:
              $ref: "#/components/schemas/DesktopLoginPollResult"
...
    DesktopLoginStartRequest:
      type: object
      properties:
        device_id:
          type: string
          nullable: true
          maxLength: 128
        device_name:
          type: string
          nullable: true
          maxLength: 128
    DesktopLoginStartResult:
      type: object
      required:
        - session_token
        - poll_token
        - authorize_url
        - verification_code
        - expires_at
        - poll_interval_seconds
      properties:
        session_token:
          type: string
          description: Included in authorize_url and safe to send to the browser.
        poll_token:
          type: string
          description: Desktop-private secret. Never display, log, or append to authorize_url.
        authorize_url:
          type: string
          description: Browser URL containing session_token and verification_code only.
        verification_code:
          type: string
          pattern: "^\\d{6}$"
        expires_at:
          type: string
          format: date-time
        poll_interval_seconds:
          type: integer
          minimum: 1
    DesktopLoginCompleteRequest:
      type: object
      required: [session_token]
      properties:
        session_token:
          type: string
    DesktopLoginCompleteResult:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [completed]
    DesktopLoginPollRequest:
      type: object
      required: [session_token, poll_token]
      properties:
        session_token:
          type: string
        poll_token:
          type: string
    DesktopLoginPollResult:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [pending, completed]
        access_token:
          type: string
          nullable: true
        refresh_token:
          type: string
          nullable: true
        expires_in:
          type: integer
          nullable: true
        token_type:
          type: string
          nullable: true
          enum: [Bearer]
        user:
          $ref: "#/components/schemas/DesktopLoginUser"
          nullable: true
          description: Present only when status=completed. Must not contain tokens, API keys or other secrets.
    DesktopLoginUser:
      type: object
      required: [id, username, email, display_name, role]
      additionalProperties: false
      properties:
        id:
          type: string
          description: Opaque user identifier; numeric backend IDs must be stringified before entering the desktop contract.
        username:
          type: string
        email:
          type: string
          format: email
        display_name:
          type: string
        role:
          type: string